requirements.txt
flask
gunicorn
apscheduler
playwright
werkzeug
requests
pymupdf
pdfplumber
python-dotenv
